Update README.md
Perform exploratory data analysis on the netflix_data.csv data to understand more about movies from the 1990s decade.
What was the most frequent movie duration in the 1990s? Save an approximate answer as an integer called duration.
A movie is considered short if it is less than 90 minutes. Count the number of short action movies released in the 1990s and save this integer as short_movie_count.